backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# LangChain / LangGraph
langchain==0.3.7
langchain-core==0.3.19
langchain-groq==0.2.1
langgraph==0.2.53

# Embeddings (local, no API key needed)
sentence-transformers==3.3.1

# Supabase
supabase==2.10.0

# Git
gitpython==3.1.43

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
httpx>=0.26,<0.28
